requirement.txt
streamlit
pillow
opencv-python-headless
easyocr
pyzbar
pandas
python-dateutil
thefuzz
rapidfuzz
